Provision SSL keystores for Kafka Brokers
This commit adds support for creating a keypair on each Kafka Broker using the supplied
Kafka CA certificate and key automatically with Puppet.

Whilst this model does mean delegating the CA key signing rights to the Kafka server, it
allows Puppet to run fully autonomously without needing a sysadmin to manually sign
systems to bring them online - important in an ever changing cloud environment.

We set a default keystore password of "password". A password is required with Java
keystores whether you want one or not, but the value can be changed via Hiera to a
more secure one if you consider keystore passwords to be a useful thing.

The CA certificate and key must be bootstrapped elsewhere and stored in Hiera. Instructions
are in the manifest for doing so.