Allocate Kafka UID/GID from the system service range, not the interactive range
Currently when the Kafka user and group are created, they take the UID/GID of
the latest human user, +1. This puts it in the 1000+ range, which results in
a future clash when we add more human users, since random IDs in the range will
be taken.

This change ensures we create them as system users/groups. This uses a
different range and avoids future clashes.

Existing systems with existing users will not be impacted.